feat: compiler versions list caching
also extracted the compiler versions switcher component to its own file as it makes sense to do so
