+++ packages/app/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@fontsource/fira-code": "^5.0.15",
     "@fontsource/montserrat": "^5.0.15",
+    "@isaacs/ttlcache": "^1.4.1",
     "@kobalte/core": "^0.11.2",
     "@solid-primitives/media": "^2.2.5",
     "@solid-primitives/scheduled": "^1.4.1",
