+++ packages/app/src/components/Settings.tsx
@@ -1,25 +1,8 @@
-import {
-  type Accessor,
-  For,
-  createMemo,
-  createResource,
-  createSignal,
-  onMount,
-  useTransition,
-  Show,
-  Suspense,
-  createEffect,
-} from "solid-js";
+import { onMount } from "solid-js";
 
 import { VsChromeClose } from "solid-icons/vs";
-import { Button, Dialog } from "@kobalte/core";
-import {
-  SegmentedButton,
-  Separator,
-  TextField,
-  ToggleButton,
-  ToggleField,
-} from "./ui-kit";
+import { Dialog } from "@kobalte/core";
+import { Separator, TextField, ToggleButton, ToggleField } from "./ui-kit";
 import {
   parsePosition,
   setParsePosition,
@@ -35,14 +18,9 @@ import {
   setTranformCompact,
   transformResultScopedSlot,
   setTransformResultScopedSlot,
-  setCurrentCompilerVersion,
   setHasCompilerVersionChangeBeenHandled,
 } from "~/lib/stores";
-// import useSWR from "solid-swr";
-import { fetchAllCompilerVersions } from "~/lib/compiler/fetch";
-import { getCompilerVersionsByType } from "~/lib/compiler/utils";
-import { setCompiler } from "~/lib/compiler";
-import { getDefaultCompilerVersionToLoad } from "~/lib/compiler/module";
+import { CompilerVersionSwitcher } from "./CompilerVersionsSwitcher";
 
 export type SettingsSectionProps = {
   closeModal: () => void;
@@ -86,7 +64,7 @@ export default function SettingsSection(props: SettingsSectionProps) {
         >
           <div class="mt-10 flex h-full flex-col space-y-10">
             {/* display the version changer */}
-            <VersionSwitcher closeModal={props.closeModal} />
+            <CompilerVersionSwitcher closeModal={props.closeModal} />
             <Separator />
             {/* display all the `Parse` options */}
             <div>
@@ -200,197 +178,6 @@ export default function SettingsSection(props: SettingsSectionProps) {
           </div>
         </Dialog.Description>
       </Dialog.Content>
-    </div>
-  );
-}
-
-function createCompilerChangeHandler(
-  afterCompilerChange: () => Promise<void> | void,
-) {
-  return (version: string) => {
-    // setHasCompilerVersionChangeBeenHandled(false);
-    setCompiler(version).then(async ({ status }) => {
-      if (status === "failure") {
-        const { compilerVersionToLoad: fallbackCompilerVersion } =
-          await getDefaultCompilerVersionToLoad();
-        // TODO: display a toast to the user
-        alert(
-          `An error occured while loading the compiler, falling back to v${fallbackCompilerVersion}`,
-        );
-        // TODO: handle failure here too
-        await setCompiler(fallbackCompilerVersion);
-        version = fallbackCompilerVersion;
-
-        if (!version) {
-          throw new Error("No fallback compiler version was found");
-        }
-      }
-      setCurrentCompilerVersion(version);
-      // setHasCompilerVersionChangeBeenHandled(true);
-      await afterCompilerChange();
-    });
-  };
-}
-
-type VersionSwitcherProps = SettingsSectionProps;
-type FetcherReturnType = Awaited<ReturnType<typeof fetchAllCompilerVersions>>;
-function compilerVersionFetcher(
-  _: boolean,
-  info: {
-    value: FetcherReturnType | undefined;
-    refetching: boolean;
-  },
-) {
-  if (info.refetching) {
-    console.log("Refetching...");
-    return fetchAllCompilerVersions({ noCache: true });
-  }
-  return fetchAllCompilerVersions();
-}
-
-function VersionSwitcher(props: VersionSwitcherProps) {
-  const STEP = 5;
-  const INITIAL_NUMBER_OF_VERSIONS_TO_DISPLAY = 10;
-  const [versionsType, setVersionsType] = createSignal<
-    "Preview" | "Production"
-  >("Production");
-  const [
-    numberOfProductionVersionsToDisplay,
-    setNumberOfProductionVersionsToDisplay,
-  ] = createSignal(INITIAL_NUMBER_OF_VERSIONS_TO_DISPLAY);
-  const [
-    numberOfPreviewVersionsToDisplay,
-    setNumberOfPreviewVersionsToDisplay,
-  ] = createSignal(INITIAL_NUMBER_OF_VERSIONS_TO_DISPLAY);
-  const [categorizedCompilerVersions, setCategorizedCompilerVersions] =
-    createSignal<ReturnType<typeof getCompilerVersionsByType>>();
-
-  const [pending, startTransition] = useTransition();
-
-  const updateNumberOfProductionVersionsToDisplay = (number: number) =>
-    startTransition(() => {
-      setNumberOfProductionVersionsToDisplay(number);
-    });
-
-  const updateNumberOfPreviewVersionsToDisplay = (number: number) =>
-    startTransition(() => {
-      setNumberOfPreviewVersionsToDisplay(number);
-    });
-
-  const handleShowMore = () => {
-    if (versionsType() === "Preview") {
-      // don't show more than the total number of preview versions
-      setNumberOfPreviewVersionsToDisplay(
-        Math.min(
-          numberOfPreviewVersionsToDisplay() + STEP,
-          categorizedCompilerVersions()!.previewVersions?.length,
-        ),
-      );
-    } else {
-      setNumberOfProductionVersionsToDisplay(
-        Math.min(
-          numberOfProductionVersionsToDisplay() + STEP,
-          categorizedCompilerVersions()!.productionVersions.length,
-        ),
-      );
-    }
-  };
-
-  const handleCompilerVersionChange = createCompilerChangeHandler(
-    props.closeModal,
-  );
-  return (
-    <div>
-      <div class="flex flex-row items-center justify-between">
-        <h3 class="text-md font-semibold leading-none text-white">Versions</h3>
-        <SegmentedButton
-          activeClass="bg-accent-2 !text-primary !font-semibold"
-          options={["Preview", "Production"]}
-          defaultActive={/*@once*/ versionsType()}
-          handleOptionChange={setVersionsType}
-        />
-      </div>
-      <Suspense fallback={<div>Loading...</div>}>
-        <VersionsList
-          setCategorizedCompilerVersions={setCategorizedCompilerVersions}
-          versionsType={versionsType}
-          numberOfProductionVersionsToDisplay={
-            numberOfProductionVersionsToDisplay
-          }
-          numberOfPreviewVersionsToDisplay={numberOfPreviewVersionsToDisplay}
-          compilerVersionChangeHandler={handleCompilerVersionChange}
-        />
-      </Suspense>
-      <div class="flex items-center justify-center">
-        <Show when={typeof categorizedCompilerVersions() !== "undefined"}>
-          <Button.Root
-            onClick={handleShowMore}
-            class="bg-zinc-700 px-3 py-2 text-zinc-50"
-          >
-            Show more
-          </Button.Root>
-        </Show>
-      </div>
-    </div>
-  );
-}
-
-type VersionsListProps = {
-  versionsType: () => "Preview" | "Production";
-  compilerVersionChangeHandler: (version: string) => void;
-  numberOfProductionVersionsToDisplay: Accessor<number>;
-  numberOfPreviewVersionsToDisplay: Accessor<number>;
-  setCategorizedCompilerVersions: (value: any) => void;
-};
-function VersionsList(props: VersionsListProps) {
-  const [allCompilerVersions, { refetch }] = createResource(
-    compilerVersionFetcher,
-  );
-  setTimeout(refetch, 0);
-
-  const categorizedCompilerVersions = createMemo(() => {
-    return getCompilerVersionsByType(allCompilerVersions() ?? []);
-  });
-
-  createEffect(() => {
-    props.setCategorizedCompilerVersions(categorizedCompilerVersions());
-  });
-
-  const productionVersionsToDisplay = createMemo(() => {
-    return categorizedCompilerVersions().productionVersions.slice(
-      0,
-      props.numberOfProductionVersionsToDisplay(),
-    );
-  });
-
-  const previewVersionsToDisplay = createMemo(() => {
-    return categorizedCompilerVersions().previewVersions.slice(
-      0,
-      props.numberOfPreviewVersionsToDisplay(),
-    );
-  });
-
-  const versionsToDisplay = createMemo(() => {
-    if (props.versionsType() === "Preview") {
-      return previewVersionsToDisplay();
-    }
-    return productionVersionsToDisplay();
-  });
-
-  return (
-    <div class="mb-10 mt-10 flex h-56 flex-wrap gap-x-8 gap-y-8 overflow-y-auto">
-      <For each={versionsToDisplay()}>
-        {(version) => {
-          return (
-            <Button.Root
-              onClick={() => props.compilerVersionChangeHandler(version)}
-              class="bg-[#222] px-3 py-2 text-zinc-100"
-            >
-              {version}
-            </Button.Root>
-          );
-        }}
-      </For>
     </div>
   );
 }
